requirements.txt
# Document Automation & Data Harmonization PoC
# AI-Powered Solution for Senior AI Engineer Interview
# Python 3.10+

# =============================================================================
# CORE AI/ML - Transformer Models
# =============================================================================

# HuggingFace Transformers - State-of-the-art models
transformers>=4.36.0

# PyTorch - Deep learning framework
torch>=2.0.0
torchvision>=0.15.0

# PEFT - Parameter-Efficient Fine-Tuning (LoRA, etc.)
peft>=0.7.0

# Sentence Transformers - For embeddings
sentence-transformers>=2.2.0

# =============================================================================
# LLM APIs - For complex reasoning fallback
# =============================================================================

openai>=1.40.0
# anthropic>=0.18.0  # Optional: Claude API

# =============================================================================
# DOCUMENT PARSING - Local extraction
# =============================================================================

pdfplumber>=0.10.0      # PDF table extraction
pandas>=2.0.0           # DataFrames
openpyxl>=3.1.0         # Excel files
python-docx>=1.1.0      # Word documents
PyMuPDF>=1.23.0         # PDF to image conversion (fitz)
pdf2image>=1.16.0       # PDF rendering (requires poppler)

# =============================================================================
# OCR - Optical Character Recognition
# =============================================================================

pytesseract>=0.3.10     # Tesseract OCR wrapper
# paddleocr>=2.7.0      # Optional: PaddleOCR (better accuracy)

# =============================================================================
# IMAGE PROCESSING
# =============================================================================

Pillow>=10.0.0          # Image handling
opencv-python>=4.8.0    # Computer vision operations
scipy>=1.11.0           # Scientific computing

# =============================================================================
# DATA VALIDATION & SCHEMAS
# =============================================================================

pydantic>=2.0.0         # Type-safe schemas

# =============================================================================
# MACHINE LEARNING UTILITIES
# =============================================================================

scikit-learn>=1.3.0     # ML utilities (clustering, etc.)
numpy>=1.24.0           # Numerical computing

# =============================================================================
# CONFIGURATION & UTILITIES
# =============================================================================

python-dotenv>=1.0.0    # Environment variables
python-dateutil>=2.8.0  # Date parsing
tqdm>=4.66.0            # Progress bars

# =============================================================================
# OPTIONAL: GPU Acceleration
# =============================================================================
# Uncomment for CUDA support:
# torch --index-url https://download.pytorch.org/whl/cu118

# =============================================================================
# OPTIONAL: Additional Models
# =============================================================================
# timm>=0.9.0           # PyTorch Image Models
# safetensors>=0.4.0    # Fast model loading

# =============================================================================
# DEVELOPMENT & TESTING
# =============================================================================
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# ruff>=0.1.0
